% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fun.R
\name{my_cov}
\alias{my_cov}
\title{Test Function}
\usage{
my_cov(x, y)
}
\arguments{
\item{x, y}{The values to compute
the covariance.}
}
\value{
The covariance.
}
\description{
A description
}
\details{
Details of the function
}
